windows: make flowtracking threadsafe in windows, and some code cleanup
